% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_distance.R
\name{add_rm_type}
\alias{add_rm_type}
\title{Relabel type of raw material}
\usage{
add_rm_type(x)
}
\arguments{
\item{x}{\code{simple features} object.}
}
\value{
\code{simple features} object.
}
\description{
Relabel type of raw material
}
