% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_distance.R
\name{count_dist_rel_rm}
\alias{count_dist_rel_rm}
\title{Count distance to period-relevant raw material sources}
\usage{
count_dist_rel_rm(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Count distance to period-relevant raw material sources
}
