% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_distance.R
\name{dist_to_rm}
\alias{dist_to_rm}
\title{Distance to a source of a raw material}
\usage{
dist_to_rm(from, to, rm, id = "id")
}
\arguments{
\item{from}{\code{sf} object of geometry type \code{POINTS} in projected
CRS (settlements).}

\item{to}{\code{sf} object of geometry type \code{POINT} or \code{LINESTRING}
in projected CRS (raw material locations).
Each raw material can have multiple locations, same raw materials are
identified by identical label in column specified by \code{rm} argument.}

\item{rm}{Column name of \code{to} to use as a grouping variable
(identical raw material).}

\item{id}{Column name of \code{from} containing unique identifiers.}
}
\value{
A \code{tibble} containing min, max, mean and median distance
for each raw material (level of \code{rm} column) to an observation
specified in \code{from} (settlement) is returned.
}
\description{
For simple feature geometries in projected coordinate reference system,
euclidean distance (in km) is returned between features of \code{from}
and \code{to} categorized by levels specified in \code{rm}.
}
