% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{est_kde}
\alias{est_kde}
\title{Helper function to estimate KDE}
\usage{
est_kde(x)
}
\arguments{
\item{x}{Simple features object. Input data.}
}
\value{
A long \code{tibble} with kernel density estimate for each id and period pair.
}
\description{
Helper function to estimate KDE
}
