% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem.R
\name{extract_from_dem}
\alias{extract_from_dem}
\title{Extract values from terrain model derived from DEM.}
\usage{
extract_from_dem(x, dem, model = "slope", fun = "mean")
}
\arguments{
\item{x}{Object of class \code{sf} or similar.}

\item{dem}{Digital elevation model.}

\item{model}{What model to use: slope, aspect, TPI, TRI...}

\item{fun}{Summary function.}
}
\value{
Object x with result in column \code{value}.
}
\description{
Extract values from terrain model derived from DEM.
}
