% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_distance.R
\name{filter_rel_dist_rm}
\alias{filter_rel_dist_rm}
\title{Filter period-relevant distances to raw material sources}
\usage{
filter_rel_dist_rm(x)
}
\arguments{
\item{x}{Input period data \code{tibble}.}
}
\value{
A \code{tibble}.
}
\description{
Filter period-relevant distances to raw material sources
}
