% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{hclust_variables}
\alias{hclust_variables}
\title{Hierarchical clustering on the variables}
\usage{
hclust_variables(x)
}
\arguments{
\item{x}{Nested PCA object.}
}
\value{
Nested object with hierarchical clustering results and dendrogram.
}
\description{
Hierarchical clustering on the variables
}
