% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{kdestimate}
\alias{kdestimate}
\title{Kernel density estimation for spatial point pattern}
\usage{
kdestimate(x, mask, markscol, lvls, sigma = 4000, points = TRUE)
}
\arguments{
\item{x}{\code{sf} object of geometry type \code{POINTS} in projected CRS.}

\item{mask}{Single \code{sf} \code{polygon}, the study window.}

\item{markscol}{Column name of \code{x} to use as grouping variable.}

\item{lvls}{Vector of levels of \code{markscol}.}

\item{sigma}{Standard deviation of isotropic smoothing kernel.
Numeric value, passed to \code{spatstat::density.splitppp} function.
Scale depends on coordinate reference system, for S-JTSK (EPSG 5514)
it is in meters and defaults to 4 km.}

\item{points}{Logical, return value for points or whole surface as raster image.
Defaults to \code{TRUE}.}
}
\value{
With \code{points} set to \code{TRUE} (default) returns a tibble with
KDE estimated at given points. If set to \code{FALSE}, a tibble with KDE
values in a grid is returned to be plotted using \code{ggplot2}.
}
\description{
Wrapper around \code{spatstat} functions. Kernel density is estimated for
a point pattern in a given window for different levels of a categorical
variable.
}
