% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{mbc_classify}
\alias{mbc_classify}
\title{Classification based on model-based clustering}
\usage{
mbc_classify(x, neo)
}
\arguments{
\item{x}{Nested object, result of \code{mbc_cluster}.}
}
\value{

}
\description{
Classification based on model-based clustering
}
