% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{mbc_cluster}
\alias{mbc_cluster}
\title{Model-based clustering}
\usage{
mbc_cluster(x, prob = 0.75)
}
\arguments{
\item{x}{A nested PCA result.}

\item{prob}{A threshold of cumulative variability explained.
Only PCs that cumulatively explain at least the given value are kept.
Defaults to 0.75.}
}
\value{
A nested \code{tibble} with \code{mclust} objects and classification
for individual points.
}
\description{
Model-based clustering
}
