% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{mbc_map}
\alias{mbc_map}
\title{Plot a map of groups resulting from model based clustering}
\usage{
mbc_map(x, sf)
}
\arguments{
\item{x}{Classification resulting from \code{mbc_classify}.}

\item{neo}{}
}
\value{

}
\description{
Plot a map of groups resulting from model based clustering
}
