% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{nested_pca}
\alias{nested_pca}
\title{Nested PCA}
\usage{
nested_pca(x)
}
\arguments{
\item{x}{A \code{tibble} grouped according to chosen variables.}
}
\value{
A nested \code{tibble} with results of \code{prcomp} function.
}
\description{
Nested PCA
}
