% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuity.R
\name{plot_continuity}
\alias{plot_continuity}
\title{Plot settlement continuity}
\usage{
plot_continuity(x, y)
}
\arguments{
\item{x}{Long results \code{tibble}.}

\item{y}{A \code{tibble} with number of settlements.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot settlement continuity
}
