% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_distance.R
\name{plot_dist_lm}
\alias{plot_dist_lm}
\title{Plot linear model for distance to raw material vs settlements kde}
\usage{
plot_dist_lm(x)
}
\arguments{
\item{x}{Long results object.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot linear model for distance to raw material vs settlements kde
}
