% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_distance.R
\name{plot_dist_rel_rm}
\alias{plot_dist_rel_rm}
\title{Plot distance to generalized raw material sources.}
\usage{
plot_dist_rel_rm(x)
}
\arguments{
\item{x}{A \code{tibble}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot distance to generalized raw material sources.
}
