% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{plot_kde}
\alias{plot_kde}
\title{Helper function to plot KDE}
\usage{
plot_kde(x)
}
\arguments{
\item{x}{Simple features object. Input data.}
}
\value{
A \code{ggplot2} object.
}
\description{
Helper function to plot KDE
}
