% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_nested_pca}
\alias{plot_nested_pca}
\title{Plot nested PCA results}
\usage{
plot_nested_pca(x, var, pc = c("PC1", "PC2"))
}
\arguments{
\item{x}{Nested PCA object, output of \code{nested_pca} function.}

\item{var}{Grouping variable visualize on.}

\item{pc}{Principal components to show on axes x and y.
Defaults to PC1 and PC2.}
}
\value{
Returns a \code{patchwork} of data in selected PC space and
associated plots with rotation visualization.
}
\description{
Plot nested PCA results
}
