% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_distance.R
\name{plot_rm_reg_dist}
\alias{plot_rm_reg_dist}
\title{Plot period-relevant distance to raw material sources in selected region}
\usage{
plot_rm_reg_dist(x, region = c("m", "b"))
}
\arguments{
\item{x}{A \code{tibble}.}

\item{region}{Region abbreviation.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot period-relevant distance to raw material sources in selected region
}
