% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis.R
\name{plot_violin}
\alias{plot_violin}
\title{Summary violin/box plots}
\usage{
plot_violin(x)
}
\arguments{
\item{x}{Long result \code{tibble}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Summary violin/box plots
}
