% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_fns.R
\name{prep_dem_res}
\alias{prep_dem_res}
\title{Prepare DEM result}
\usage{
prep_dem_res(x)
}
\arguments{
\item{x}{A \code{tibble}.}
}
\value{
A long results \code{tibble}.
}
\description{
Functions with \code{prep_} prefix prepare the data into the long format for
binding with results \code{tibble}.
}
