% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_fns.R
\name{prep_dist_fenced}
\alias{prep_dist_fenced}
\title{Prepare distance to fenced settlements result}
\usage{
prep_dist_fenced(x)
}
\arguments{
\item{x}{A \code{tibble}.}
}
\value{
A long results \code{tibble}.
}
\description{
Functions with \code{prep_} prefix prepare the data into the long format for
binding with results \code{tibble}.
}
