% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_distance.R
\name{prep_dist_rel_rm}
\alias{prep_dist_rel_rm}
\title{Prepare period-relevant distance to the generic type of raw material source}
\usage{
prep_dist_rel_rm(x)
}
\arguments{
\item{x}{A \code{tibble}.}
}
\value{
A long results \code{tibble}. Raw materials generalized into
\code{Chipped} and \code{Polished} categories.
}
\description{
Functions with \code{prep_} prefix prepare the data into the long format for
binding with results \code{tibble}.
}
