% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_fns.R
\name{prep_wide_res}
\alias{prep_wide_res}
\title{Prepare wide results \code{tibble}}
\usage{
prep_wide_res(x)
}
\arguments{
\item{x}{A long results \code{tibble}.}
}
\value{
A wide results \code{tibble}.
}
\description{
Transforms long results \code{tibble} into a wide \code{tibble} with
variables in columns for different settlements adn periods.
}
