% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minions.R
\name{relabel_variable}
\alias{relabel_variable}
\title{Abbreviate variable names}
\usage{
relabel_variable(x)
}
\arguments{
\item{x}{A \code{tibble}.}
}
\value{
A \code{tibble}.
}
\description{
Abbreviate variable names
}
