% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuity.R
\name{settlement_continuity}
\alias{settlement_continuity}
\title{Derive settlement continuity across periods}
\usage{
settlement_continuity(x)
}
\arguments{
\item{x}{Input period \code{tibble}.}
}
\value{
A \code{tibble}.
}
\description{
Derive settlement continuity across periods
}
