% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{summarise_variables}
\alias{summarise_variables}
\title{Summarise variable loadings}
\usage{
summarise_variables(x, prob = 0.75)
}
\arguments{
\item{x}{Nested PCA object.}

\item{prob}{A threshold of cumulative variability explained.
Only PCs that cumulatively explain at least the given value are kept.
Defaults to 0.75.}
}
\value{
A tibble with variable loadings summarised across different PCAs.
}
\description{
Summarise variable loadings
}
