% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_distance.R
\name{tab_dist_lm}
\alias{tab_dist_lm}
\title{Calculate correlation between set. KDE and dist. to raw materials}
\usage{
tab_dist_lm(x)
}
\arguments{
\item{x}{Long results object.}
}
\value{
A \code{kable} object.
}
\description{
Calculate correlation between set. KDE and dist. to raw materials
}
