% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minions.R
\name{table_nr_set}
\alias{table_nr_set}
\title{Tabulate input data}
\usage{
table_nr_set(x)
}
\arguments{
\item{x}{Input data \code{tibble}.}
}
\value{
A \code{kable} object.
}
\description{
Tabulate input data
}
